sx1255

Module hard_registers

Semtech SX1255 Hardware Registers

This module contains structures similar to the the hardware registers as used on the SX1255, and software to serialize them to the format of the actual hardware registers. The Rust data structures differ from the actual hardware registers where the data sizes are not directly supported by Rust, for example frequencies on the SX1255 are 24 bits, and the Rust type for them is a std::primitive::u32. Nor does Rust natively support bit-fields. The HardRegisters::serialize function serializes the data to fit the actual register formats.

Some of these structures are not directly used by the programmer. Instead, the module registers defines similar data structures using more convenient data formats in the engineer’s accustomed units: dB, femtoFarad, ohm, etc.; rather than a binary encoding. Registers::write translates its data to the HardRegisters data format, serializes that to the actual data format used by the IC hardware, and uses the SPI bus to write the hardware registers.

At this writing the SX1255 data sheet was at https://semtech.my.salesforce.com/sfc/p/#E0000000JelG/a/44000000MDmE/Qs9oRoa8Sbb6mkImE9mtMh47H5LFx6KMbGcpb8L28SE All the structures and fields correspond to the data sheet names, except that in a few places I’ve attempted to make the names more readable, but they should still be recognizable.

Structs

ClockSelect

SX1255 hardware mapping of clock select and loop-back modes.

DigBridge

SX1255 hardware mapping of status bits.

Frequency

Integer frequency value. The actual frequency will be (oscillator_frequency * frequency_value) / 2^20 .

HardRegisters

Semtech SX1255 Hardware Registers

IISM

SX1255 hardware mapping of status bits.

IOMap

SX1255 hardware mapping of the 4 DIO pins.

Mode

SX1255 hardware mode register.

RxFrontend

SX1255 hardware receive front-end control register.

Status

SX1255 hardware mapping of status bits.

TxFrontend

SX1255 hardware transmit front-end control register.

Version

SX1255 hardware transmit front-end control register.

Enums

IISMMode

IOMap0

IOMap1

IOMap2

IOMap3

Functions

_stub 🔒